Correct an issue in running PHPUnit coverage
When running PHPUnit with
--coverage-html phpunit-coverage,
there was an error:
Error: Incorrect whitelist config, no code coverage will be generated.

This could be a problem with the version
of PHPUnit I'm using locally.
It's tough getting the right version of
PHPUnit.
Because we're testing PHP 5.6 in CircleCI,
we have to use 5.7.9 there,
but that doesn't work in my PhpStorm.