Only load Deprecated.php if block_row() doesn't exist
GCB Pro might have this,
in which case loading it again would cause
a PHP error.
So give it a chance to load it first,
by loading this at the priority of 11.
So at least BL users that still call
block_row() and the Repeater functions
won't have PHP errors, only deprecated warnings.