Simpler solution to dupicate name
If 2 fields have the same name,
append an empty space to one.
This is a little hacky,
but the previous commits were much more complex,
and involved the key to the field and
the field.name getting out of sync.
This seems to work fine.
Though it doesn't allow 3 names to be the same at the same time.