test: remove no field test
This test is intended to check that the editor page still loads if the
user creates a custom field with no fields.

The test passes locally but fails in Circle. Rather than continuing to
tweak it to attempt to appease Circle I opted to delete it.
It seems to have limited  value since (a) it passes right now and
(b) it doesn't seem common that a user will add a custom field
with no fields.