chore: bump 'tested up to' to 6.9
Includes temporary downgrade of version number to align with WP.org
version, which is one behind. (1.7.1 was only deployed to WPE servers.)

Required because the CI workflow to bump the 'tested up to' on the
remote SVN repo uses that version value to determine which
tags/[version] folder to update.

Version downgrade will be reverted in the next commit.